Creater header and overlay
