+++ frontend/src/App.css
@@ -75,27 +75,32 @@ h3 {
   color: #fff;
 }
 
-.user-info {
+.header-container {
   display: flex;
+  justify-content: space-between;
   align-items: center;
-  padding: 10px 20px;
-  background-color: #111;
-  border-bottom: 1px solid #1db954;
+  padding: 20px;
+  border-bottom: 1px solid #fff;
+  height: 15px;
 }
 
-.user-info h2 {
-  margin: 0;
-  font-size: 20px;
-}
-
-.user-info p {
-  margin: 5px 0 0;
-  font-size: 14px;
-  color: #aaa;
+.header-container h1 {
+  font-size: 1.5rem;
 }
 
 .globe-container {
+  position: relative;
   flex-grow: 1;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.overlay-container {
+  position: absolute;
+  height: 100%;
+  width: 100%;
+  z-index: 100;
   display: flex;
   justify-content: center;
   align-items: center;
@@ -112,4 +117,14 @@ h3 {
   flex-direction: column;
   top: 50px;
   right: 50px;
+}
+
+.artist-info {
+  display: flex;
+  align-items: center;
+}
+
+img.artist-image {
+  width: 50px;
+  margin-right: 15px;
 }
