+++ frontend/src/components/CountryInfo.js
@@ -6,8 +6,8 @@ const CountryInfo = ({ selectedCountry }) => {
       <h2>Top artists listed to by you in {selectedCountry.countryName}</h2>
       <ul>
         {selectedCountry.items.map((item, index) => (
-          <div key={index}>
-            <img src={item.image} alt=""></img>
+          <div className="artist-info" key={index}>
+            <img className="artist-image" src={item.image} alt=""></img>
             <span>{item.name}</span>
           </div>
         ))}
