+++ frontend/src/components/Dashboard.js
@@ -1,29 +1,17 @@
-import React, { useState, useEffect } from "react";
+import React, { useState } from "react";
 import GlobeComponent from "./Globe";
 import spotifyApi from "../services/spotify";
 import "../App.css";
 import CountryInfo from "./CountryInfo";
+import Header from "./Header";
+import OverlayDashboard from "./OverlayDashboard";
 
 const Dashboard = () => {
-  const [userData, setUserData] = useState(null);
   const [artistsData, setArtistsData] = useState([]);
   const [highlightedCountries, setHighlightedCountries] = useState([]);
   const [selectedCountry, setSelectedCountry] = useState(null);
   const [isRotating, setIsRotating] = useState(false);
   const [isInteractive, setIsInteractive] = useState(true);
-
-  useEffect(() => {
-    const fetchUserData = async () => {
-      try {
-        const response = await spotifyApi.fetchUserInfo();
-        setUserData(response);
-      } catch (error) {
-        console.error(error);
-      }
-    };
-
-    fetchUserData();
-  }, []);
 
   const fetchTopTracks = async () => {
     setArtistsData([]);
@@ -53,7 +41,6 @@ const Dashboard = () => {
           (artist) => artist.country === country.ISO_A2
         ),
       };
-      console.log(selectedInfo);
       setSelectedCountry(selectedInfo);
     } else {
       setSelectedCountry(null);
@@ -62,19 +49,11 @@ const Dashboard = () => {
 
   return (
     <div className="dashboard">
-      <div className="user-info">
-        {/* <img src={userData?.}></img> */}
-        <div>
-          <h2>{userData?.display_name}</h2>
-          <p>{userData?.email}</p>
-        </div>
-        <p>
-          You have listened to artists from {highlightedCountries.length}{" "}
-          countries. Click on a country to see your top songs there.
-        </p>
-        <button onClick={fetchTopTracks}>Fetch My Top Tracks</button>
-      </div>
+      <Header />
       <div className="globe-container">
+        {!artistsData.length && (
+          <OverlayDashboard onFetch={fetchTopTracks} isLoading={isRotating} />
+        )}
         <GlobeComponent
           highlightedCountries={highlightedCountries}
           isRotating={isRotating}
