+++ frontend/src/components/OverlayDashboard.js
@@ -1,0 +1,17 @@
+import React from "react";
+
+const OverlayDashboard = ({ onFetch, isLoading }) => {
+  return (
+    <div className="overlay-container">
+      {isLoading ? (
+        <div className="loader">Loading.....</div>
+      ) : (
+        <button className="fetch-button" onClick={onFetch}>
+          Fetch Top Tracks
+        </button>
+      )}
+    </div>
+  );
+};
+
+export default OverlayDashboard;
